プロファイルごとにテストクラスを分割
